requirements.txt
cryptography==3.4.7
python-telegram-bot
openai
